pyproject.toml
[project]
name = "agentic-workflow"
version = "0.1.0"
description = "AI Agent Pipeline with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.23",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.21",
    "langchain-groq>=0.1.3",
    "langgraph>=0.4.8",
    "streamlit>=1.31.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pytest>=7.4.3",
    "langchain-qdrant>=0.2.0",
    "langchain-voyageai>=0.1.4",
    "pypdf>=5.4.0",
    "langchain-cohere>=0.4.3",
    "pytest-cov>=6.1.0",
    "langsmith>=0.3.24",
    "tavily-python>=0.7.5",
    "langchain-tavily>=0.2.1",
    "langchain-experimental>=0.3.4",
    "ipython>=8.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
